requirements.txt
numpy>=1.20.0
matplotlib>=3.4.0
opencv-python>=4.5.0
scipy>=1.7.0
pandas>=1.3.0
llvmlite>=0.39.0
numba>=0.56.0
ttkbootstrap>=1.10.1
Pillow>=9.0.0  # For icon creation